requirements.txt
streamlit
pandas
openai==0.28.0
google-generativeai
python-docx
Pillow
reportlab
langchain
faiss-cpu
tqdm
docx2txt
PyMuPDF
unstructured
sentence-transformers==2.6.1
transformers==4.40.0
python-pptx
lxml
beautifulsoup4
langchain_community
